Created search field
