+++ src/components/Header.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 
@@ -8,11 +8,6 @@ import { logout } from '../redux/actions/user';
 
 const Header = () => {
   const dispatch = useDispatch();
-  const [search, setSearch] = useState('');
-
-  const handleChangeSearch = ({ target }) => {
-    setSearch(target.value);
-  };
 
   const handleLogout = () => {
     dispatch(logout());
@@ -24,17 +19,6 @@ const Header = () => {
         Book Shop
       </Link>
       <div className="header__info">
-        <form className="header__form">
-          <input
-            className="header__form-input"
-            type="text"
-            value={search}
-            onChange={handleChangeSearch}
-          />
-          <button className="header__form-btn" type="button">
-            Search
-          </button>
-        </form>
         <div className="header__cart" onClick={handleLogout} aria-hidden>
           <img className="header__cart-image" src={logoutPng} alt="logout" />
           <span className="header__cart-label">Logout</span>
