+++ src/scss/components/header.scss
@@ -20,41 +20,6 @@
     display: flex;
   }
 
-  &__form {
-    display: flex;
-    align-items: center;
-    height: 40px;
-    margin-right: 50px;
-    background-color: #ffffff;
-  }
-
-  &__form-input {
-    font-weight: normal;
-    font-size: 16px;
-    width: 100%;
-    height: inherit;
-    padding-left: 14px;
-    padding-right: 14px;
-    border: 1px solid $black;
-    outline: none;
-  }
-
-  &__form-btn {
-    cursor: pointer;
-    padding: 10px 25px;
-    background-color: $black;
-    font-size: 15px;
-    height: inherit;
-    color: $white;
-    outline: none;
-    border: 1px solid $black;
-
-    &:hover {
-      transition: 0.5s;
-      opacity: 0.5;
-    }
-  }
-
   &__cart {
     cursor: pointer;
     margin-right: 25px;
