+++ src/scss/components/nav.scss
@@ -1,0 +1,14 @@
+.nav {
+  padding: 30px 60px;
+  border-bottom: 1px solid #eeeeee;
+
+  &__form-input {
+    font-size: 16px;
+    max-width: 300px;
+    width: 100%;
+    height: 40px;
+    padding-left: 14px;
+    padding-right: 14px;
+    border: 1px solid $black;
+  }
+}
