+++ src/scss/index.scss
@@ -1,5 +1,6 @@
 @import '_colors';
 @import 'components/header';
+@import 'components/nav';
 @import 'pages/login';
 @import 'pages/detail';
 @import 'pages/cart';
